Internal Documentation

Documentation for internal SeisMain.jl interface

Internals

SeisMain.InitSeisHeader — Function

InitSeisHeader

Initialize a variable of composite type Header corresponding to the data header in seis format. All the fields are initialized to 0.0. Type ?SeisMain.Header for a detail of the fields included.

SeisMain.GrabHeader — Function

GrabHeader(stream,j)

Extract the header from trace j of a file in seis format.

Arguments

stream::IOStream: data file in seis format

j::Integer: Trace number

SeisMain.PutHeader — Function

PutHeader(stream,j)

Write the header of trace j to a file in seis format.

Arguments

stream::IOStream: data file in seis format

j::Integer: Trace number

SeisMain.BitsToHeader — Function

BitsToHeader(h_in)

Reinterpret an input header to its correct type. The function outputs a variable of composite type Header.

SeisMain.HeaderToBits — Function

HeaderToBits(h_in)

Reinterpret a composite type variable of type Header to 32 bits The function outputs an Vector with 31 elements of type Header32Bit.

SeisMain.GetNumTraces — Function

GetNumTraces(file)

Calculate the total number of traces in a data file with seis format

SeisMain.ParseHeaderName — Function

ParseHeaderName(file::AbstractString)

Parse the name of the header file from an extent file to a String variable.

SeisMain.ParseDataName — Function

ParseDataName(file::AbstractString)

Parse the name of the data file from an extent file to a String variable.

SeisMain.ParseDataFormat — Function

ParseDataFormat(file::AbstractString)

Parse the format of the data type in the extent file to a String variable.

SeisMain.ParseDataESize — Function

ParseDataESize(file::AbstractString)

Parse the esize in the extent file to a Int32 variable.

SeisMain.ExtractHeader — Function

ExtractHeader(h::Array{Header,1},key::AbstractString)

Extracts the values of a field in the header. The output is a vector of the same type as the field.

SeisMain.ReadTextHeader — Function

ReadTextHeader(filein)

Reads the extent file of data in seis format. The result outputs in a variable of composite type Extent. Type ?SeisMain.Extent for a detail on the included fields.

SeisMain.WriteTextHeader — Function

WriteTextHeader(filename,extent,format,esize,filename_d,filename_h)

Writes the extent file, corresponding to the seis format, to the file filename.

Arguments

extent: variable of type Extent to write to file

format: format of the data type in file @data@

esize: esize

filename_d: path of the @data@ file as a String variable

filename_h: path of the @headers@ file a a String variable

SeisMain.InitFileHeader — Function

InitFileHeader

Initializes a variable of composite type fileHeader corresponding to the file header in SEGY format. All the fields are initialized to 0. Type ?SeisMain.fileHeader for a detail of the fields included.

SeisMain.GrabFileHeader — Function

GrabFileHeader(stream)

Extracts the file header from a standard SEGY rev0 data file, starting from byte 3200.

SeisMain.PutFileHeader — Function

PutFileHeader(stream,fh)

Writes the file header on a standard SEGY rev0 format file.

Arguments

stream::IOStream: data file in seis format

fh::fileHeader: file Header to write

SeisMain.InitSegyHeader — Function

InitSegyHeader()

Initializes a variable of composite type SegyHeader corresponding to the trace header in SEGY format. All the fields are initialized to 0. Type ?SeisMain.SegyHeader for a detail of the fields included.

SeisMain.GrabSegyHeader — Function

GrabSegyHeader(stream,swap_bytes,nt,file_header_size,j)

Extracts the header from trace j of a file in segy format. The output is of composite type SegyHeader.

Arguments

stream::IOStream: data file in segy format

swap_bytes: boolean variable. Consider it for indianness

nt: time samples.

file_header_size: Size of file header in bytes: 0 for su files, 3600 for Segy rev0 standard.

j::Integer: Trace number

SeisMain.PutSegyHeader — Function

PutSegyHeader(stream,h,nt,file_header_size,j)

Writes the header of trace j to a file in segy format.

Arguments

stream::IOStream: data file in segy format

h::SegyHeader: header variable to write

file_header_size: Size of file header in bytes: 0 for su files, 3600 for Segy rev0 standard.

j::Integer: Trace number

SeisMain.MapHeaders — Function

MapHeaders(h_in,j,map_type)

Maps header attributes from seis byte location to segy byte location in the file and backwards. The output is of type Header or SegyHeader accordingly.

Arguments

h_in: header input data

j::Integer: Trace number

map type: "SegyToSeis" or "SeisToSegy"

Base.convert — Function

convert(::Type{Float32}, ibm::IBMFloat32)

Extension of function convert in SeismicJulia to cover IBMFloat32 Floating-point format