Public

Public Documentation

Documentation for public SeisMain.jl public interface

Public interface

SeisMain.SegyToSeis — Function.

SegyToSeis(filename_in,filename_out;<keyword arguments>)

Convert SEGY or SU data to seis format. The function needs input and output filenames.

Arguments

format="segy" : Options are segy or su

swap_bytes=true : If the flag equals true, the function swaps bytes

input_type="ibm" : Options are ibm or ieee

Credits: AS, 2015

source

SeisMain.SeisRead — Function.

SeisRead(filename;<keyword arguments>)

Read seismic data from a given filename in seis format. The format is comprised of three elements:

a text file (data extent) with geometry information

a binary file containing data (@data@)

a binary file containing headers (@headers@)

Keyword arguments

group="all" : Options are all, some or gather

key=["imx","imy"]

itrace=1 : Number of trace where the function starts reading

ntrace=10000 : Total number of traces to read

Out

d: data as 2d array

h: headers as 1d array

extent: extent of the data (try fieldnames(Extent) to see the information this contains)

Example

d,h,ext = SeisRead(filename)

Credits: AS, 2015

source

SeisMain.SeisReadHeaders — Function.

SeisReadHeaders(filename;<keyword arguments>)

Read the headers of a input file in seis format

Arguments

group="all" : Options are all, some or gather

key=[]

itrace=1 : Number of trace where the function starts reading

ntrace=100 : Total number of traces to read

Example

h = SeisRead(filename)

Credits: AS, 2015

source

SeisMain.SeisWrite — Function.

SeisWrite(filename,d,h,extent;<keyword arguments>)

Write seismic data in seis format

Arguments

filename : Name of file to write/generate

d: seismic data

h::Array{Header,1}: headers as 1d array with elements of type Header

extent::Extent : extent of the data (try names(Extent) to see the information this contains)

itrace=1 : First trace number to write

Credits: AS, 2015

source

SeisMain.SeisWriteHeaders — Function.

SeisWriteHeaders(filename,h;<keyword arguments>)

Write seismic headers in seis format

Arguments

itrace=1 : First trace number to write

update_tracenum=true

Credits: AS, 2015

source

SeisMain.SeisCopy — Function.

SeisCopy(in,out)

Copy a seis input file to a seis output file. In and out should be of type AbstractString

Credits: AS, 2015

source

SeisMain.SeisHeaderInfo — Function.

SeisHeaderInfo(filename;<keyword arguments>)

Print Seis header information to screen. The input is the name of the data file

Arguments

ntrace=100000 : Number of traces to analyze

Credits: AS, 2015

source

SeisMain.SeisToSegy — Function.

SeisToSegy(filename_in,filename_out;<keyword arguments>)

Convert seis data to SU or SEGY format. The function needs input and output filenames.

Arguments

su=true : If the flag equals true, converts tu SU format, otherwise to SEGY format

Credits: AS, 2015

source

SeisMain.SeisRemove — Function.

SeisRemove(filename)

Delete a seis file (deletes the text file, binary data file, and binary header file if there is one)

Credits: AS, 2015

source