Public

Public Documentation

Documentation for public SeisMain.jl public interface

Public interface

SeisMain.SegyToSeis — Function.

SegyToSeis(filename_in,filename_out;<keyword arguments>)

Convert SEGY or SU data to seis format. The function needs input and output filenames.

Arguments

format="segy" : Options are segy or su

swap_bytes=true : If the flag equals true, the function swaps bytes

input_type="ibm" : Options are ibm or ieee

Credits: AS, 2015

source

SeisMain.SeisRead — Function.

SeisRead(filename;<keyword arguments>)

Read seismic data from a given filename in seis format. The format is comprised of three elements:

a text file (data extent) with geometry information

a binary file containing data (@data@)

a binary file containing headers (@headers@)

Keyword arguments

group="all" : Options are all, some or gather

key=["imx","imy"]

itrace=1 : Number of trace where the function starts reading

ntrace=10000 : Total number of traces to read

Out

d: data as 2d array

h: headers as 1d array

extent: extent of the data (try fieldnames(Extent) to see the information this contains)

Example

d,h,ext = SeisRead(filename)

Credits: AS, 2015

source

SeisMain.SeisReadHeaders — Function.

SeisReadHeaders(filename;<keyword arguments>)

Read the headers of a input file in seis format

Arguments

group="all" : Options are all, some or gather

key=[]

itrace=1 : Number of trace where the function starts reading

ntrace=100 : Total number of traces to read

Example

h = SeisRead(filename)

Credits: AS, 2015

source

SeisMain.SeisWrite — Function.

SeisWrite(filename,d,h,extent;<keyword arguments>)

Write seismic data in seis format

Arguments

filename : Name of file to write/generate

d: seismic data

h::Array{Header,1}: headers as 1d array with elements of type Header

extent::Extent : extent of the data (try names(Extent) to see the information this contains)

itrace=1 : First trace number to write

Credits: AS, 2015

source

SeisMain.SeisWriteHeaders — Function.

SeisWriteHeaders(filename,h;<keyword arguments>)

Write seismic headers in seis format

Arguments

itrace=1 : First trace number to write

update_tracenum=true

Credits: AS, 2015

source

SeisMain.SeisCopy — Function.

SeisCopy(in,out)

Copy a seis input file to a seis output file. In and out should be of type AbstractString

Credits: AS, 2015

source

SeisMain.SeisHeaderInfo — Function.

SeisHeaderInfo(filename;<keyword arguments>)

Print Seis header information to screen. The input is the name of the data file

Arguments

ntrace=100000 : Number of traces to analyze

Credits: AS, 2015

source

SeisMain.SeisToSegy — Function.

SeisToSegy(filename_in,filename_out;<keyword arguments>)

Convert seis data to SU or SEGY format. The function needs input and output filenames.

Arguments

su=true : If the flag equals true, converts tu SU format, otherwise to SEGY format

Credits: AS, 2015

source

SeisMain.SeisRemove — Function.

SeisRemove(filename)

Delete a seis file (deletes the text file, binary data file, and binary header file if there is one)

Credits: AS, 2015

source

SeisMain.SeisWindow — Function.

SeisWindow(in,out;<keyword arguments>)

Window a seis file using header words.

Arguments

in::String: filename of input

out::String: filename of output

Keyword arguments

key=[]

minval=[]

maxval=[]

note that windowing along the time axis is achieved by using the key "t".

Credits: AS, 2015

source

SeisMain.SeisBinData — Function.

SeisBinData(in,out; <keyword arguments>)

Sequentially bin seismic data using already binned trace headers. Input arguments should be consistent with SeisBinHeaders input arguments. See also: SeisBinHeaders

Arguments

in: filename of input, irregularly sampled data

out: filename of output, regularly sampled data

Keyword arguments

style="sxsygxgy": bin style. Options: "mxmyhxhy","mxmyhaz","sxsyhxhy","gxgyhxhy","sxsyhaz","gxgyhaz"

ang=90: inline direction measured in degrees CC from East

gamma=1: vp/vs ratio for PS Asymptotic Conversion Point gathers (use gamma=1 for PP data)

osx=0,osy=0,ogx=0,ogy=0 : origin for source and receiver coordinate system

omx=0,omy=0,ohx=0,ohy=0: origin for midpoint and offset coordinate system

oaz=0,oh=0 : origin for azimuth and offset coordinate system

dsx=1,dsy=1,dgx=1,dgy=1: source and receiver step-size

dmx=1,dmy=1,dhx=1,dhy=1: midpoint and offset step-size

dh=1,daz=1: offset and azimuth step-size

min_isx=0,max_isx=0,min_isy=0,max_isy=0: grid extreme values for sources

min_igx=0,max_igx=0,min_igy=0,max_igy=0: grid extreme values for receivers

min_imx=0,max_imx=0,min_imy=0,max_imy=0: grid extreme values for midpoints

min_ihx=0,max_ihx=0,min_ihy=0,max_ihy=0: grid extreme values for offsets

min_ih=0,max_ih=0,min_iaz=0,max_iaz=0: grid extreme values for azimuth and offset

ntrace=10000: maximum number of traces processed at a time

Output

In file out, the binned data is saved.

Credits: Aaron Stanton, 2017

source

SeisMain.SeisPatch — Function.

  SeisPatch(in,out;<keyword arguments>)

Creates overlapping 5d patches from a 5d volume

Arguments

in::String: input filename (data should have grid information in headers)

out::String: prefix for output filenames

Keyword arguments

style="sxsygxgy": bin style. Options: "mxmyhxhy","mxmyhaz","sxsyhxhy","gxgyhxhy","sxsyhaz","gxgyhaz"

min_isx=0,max_isx=0,min_isy=0,max_isy=0: grid extreme values for sources

min_igx=0,max_igx=0,min_igy=0,max_igy=0: grid extreme values for receivers

min_imx=0,max_imx=0,min_imy=0,max_imy=0: grid extreme values for midpoints

min_ihx=0,max_ihx=0,min_ihy=0,max_ihy=0: grid extreme values for offsets

min_ih=0,max_ih=0,min_iaz=0,max_iaz=0: grid extreme values for azimuth and offset

it_WL=9e9,it_WO=0 : length and overlapping samples in time patches

ix1_WL=9e9,ix1_WO=0:length and overlapping samples in first space dimension

ix2_WL=9e9,ix2_WO=0,ix3_WL=9e9,ix3_WO=0,ix4_WL=9e9,ix4_WO=0

Output

filename,npatch: String Array with the file name of the data patches, number of patches created

Credits: A. Stanton

source

SeisMain.SeisUnPatch — Function.

SeisUnPatch(in,out;<keyword arguments>)

Reconstruct a 5D data volume from a set of 5D data patches.

Arguments

in::Array{String,1}: array containing filename of patches

out::String: filename for reconstructed volume

Keyword arguments

style="sxsygxgy": bin style. Options: "mxmyhxhy","mxmyhaz","sxsyhxhy","gxgyhxhy","sxsyhaz","gxgyhaz"

min_isx=0,max_isx=0,min_isy=0,max_isy=0: grid extreme values for sources

min_igx=0,max_igx=0,min_igy=0,max_igy=0: grid extreme values for receivers

min_imx=0,max_imx=0,min_imy=0,max_imy=0: grid extreme values for midpoints

min_ihx=0,max_ihx=0,min_ihy=0,max_ihy=0: grid extreme values for offsets

min_ih=0,max_ih=0,min_iaz=0,max_iaz=0: grid extreme values for azimuth and offset

it_WL=9e9,it_WO=0 : length and overlapping samples in time patches

ix1_WL=9e9,ix1_WO=0:length and overlapping samples in first space dimension

ix2_WL=9e9,ix2_WO=0,ix3_WL=9e9,ix3_WO=0,ix4_WL=9e9,ix4_WO=0

nt=0: time samples of reconstructed cube

ang=90: inline direction measured in degrees CC from East

gamma=1: vp/vs ratio for PS Asymptotic Conversion Point gathers (use gamma=1 for PP data)

osx=0,osy=0,ogx=0,ogy=0 : origin for source and receiver coordinate system

omx=0,omy=0,ohx=0,ohy=0: origin for midpoint and offset coordinate system

oaz=0,oh=0 : origin for azimuth and offset coordinate system

dsx=1,dsy=1,dgx=1,dgy=1: source and receiver step-size

dmx=1,dmy=1,dhx=1,dhy=1: midpoint and offset step-size

dh=1,daz=1: offset and azimuth step-size

Output

In file out, the 5D reconstructed volume is created.

Credits: A. Stanton, F Carozzi, 2017

source

SeisMain.SeisGeometry — Function.

SeisGeometry(in;<keyword arguments>)

Update headers with geometry information. Offsets and azimuths are calculated from source and receivers coordinates.

Arguments

in: input filename

Keyword arguments

ang=90: inline direction measured in degrees CC from East

gamma=1: vp/vs ratio for PS Asymptotic Conversion Point gathers (use gamma=1 for PP data)

osx=0,osy=0,ogx=0,ogy=0 : origin for source and receiver coordinate system

omx=0,omy=0,ohx=0,ohy=0: origin for midpoint and offset coordinate system

oaz=0,oh=0 : origin for azimuth and offset coordinate system

dsx=1,dsy=1,dgx=1,dgy=1: source and receiver step-size

dmx=1,dmy=1,dhx=1,dhy=1: midpoint and offset step-size

dh=1,daz=1: offset and azimuth step-size

Outputs

the @headers@ file is updated with the following information:

hx,hy,h,az,mx,my : calculated offset, azimuth and midpoint

isx,isy,igx,igy,imx,imy,ihx,ihy,ih,iaz: calculated grid nodes for source and receiver position and midpoint, offset and azimuth.

Credits: A. Stanton, 2017

source

SeisMain.SeisBinHeaders — Function.

SeisBinHeaders(in,out; <keyword arguments>)

Sequentially bin seismic headers using the available grid information.

Keyword arguments should be consistent with SeisGeometry keyword arguments.

Arguments

in: filename of input, irregularly sampled data

out: filename of output, regularly sampled data

Keyword arguments

style="sxsygxgy": bin style. Options: "mxmyhxhy","mxmyhaz","sxsyhxhy","gxgyhxhy","sxsyhaz","gxgyhaz"

ang=90: inline direction measured in degrees CC from East

gamma=1: vp/vs ratio for PS Asymptotic Conversion Point gathers (use gamma=1 for PP data)

osx=0,osy=0,ogx=0,ogy=0 : origin for source and receiver coordinate system

omx=0,omy=0,ohx=0,ohy=0: origin for midpoint and offset coordinate system

oaz=0,oh=0 : origin for azimuth and offset coordinate system

dsx=1,dsy=1,dgx=1,dgy=1: source and receiver step-size

dmx=1,dmy=1,dhx=1,dhy=1: midpoint and offset step-size

dh=1,daz=1: offset and azimuth step-size

min_isx=0,max_isx=0,min_isy=0,max_isy=0: grid extreme values for sources

min_igx=0,max_igx=0,min_igy=0,max_igy=0: grid extreme values for receivers

min_imx=0,max_imx=0,min_imy=0,max_imy=0: grid extreme values for midpoints

min_ihx=0,max_ihx=0,min_ihy=0,max_ihy=0: grid extreme values for offsets

min_ih=0,max_ih=0,min_iaz=0,max_iaz=0: grid extreme values for azimuth and offset

ntrace=10000: maximum number of traces processed at a time

Output

In file out@headers@, binned headers are saved.

Credits: Aaron Stanton,2017

source

Missing docstring.

Missing docstring for SeisBinData. Check Documenter's build log for details.